Refactor Reducer into separate components
