+++ app/src/reducer/data.js
@@ -1,0 +1,11 @@
+import { Map } from 'immutable';
+
+const defaultState = Map({});
+
+export default function data(state = defaultState, action) {
+  switch (action.type) {
+    
+    default:
+      return state
+  }
+}
