+++ app/src/utils/messages.js
@@ -1,4 +1,7 @@
-export function parseError(response) {
+import { fromJS } from 'immutable';
+
+
+function parseError(response) {
   let errorList = [];
 
   for (var key in response) {
@@ -6,4 +9,18 @@ export function parseError(response) {
   }
 
   return errorList;
+}
+
+export function createMessage(type, msg, response) {
+  var message = { type, msg };
+
+  if (response) {
+  	if (response.network_error) {
+      message.msg = response.network_error;
+  	} else {
+  	  message.details = parseError(response);
+  	}
+  }
+
+  return fromJS(message);
 }
