+++ app/test/store/test_store.js
@@ -1,26 +1,26 @@
-import {Map, fromJS} from 'immutable';
+import {fromJS} from 'immutable';
 import {expect} from 'chai';
 
 import store from '../../src/store';
-import { INITIAL_STATE } from '../../src/reducer'
 
 describe('store', () => {
 
   it('is a Redux store configured with the correct reducer', () => {
-    expect(store.getState()).to.equal(INITIAL_STATE);
-
-    store.dispatch({
-      type: 'REQUEST_START'
+    expect(store.getState()).to.deep.equal({
+      user: fromJS({ }),
+      messages: fromJS({
+              requestsPending: 0,
+              userFeedback: [ ]
+            }),
+      data: fromJS({ })
     });
 
-    expect(store.getState()).to.equal(fromJS({
-      user: {},
-      messages: {
-        requestsPending: 1,
-        userFeedback: []
-      },
-      data: {}
-    }));
+    store.dispatch({ type: 'REQUEST_START' });
+
+    const state = store.getState();
+    expect(state).to.have.property('user');
+    expect(state).to.have.property('data');
+    expect(state.messages.get('requestsPending')).to.equal(1);
   });
 
 });
